Rename agora to pinax-forums and move to Pinax organization
agora was developed a while ago for Eldarion clients who permitted us
open sourcing it. We open sourced it under the name agora as Eldarion
open source but from the start it was considered part of Pinax.

Since we have yet to release a package and are preparing to do so, it
made sense to relocate the package under the Pinax organization and
rename it to something so it’s clearly a Pinax app.